Implemented bank playment. PCC Left out.

+++ backend/SEP/BankService/Interfaces/IAccountService.cs
@@ -5,7 +5,7 @@
         Task<int> GetAccountId(string acquirerAccountNumber);
         Task <string> GetAccountNumberByMerchant(int merchantId);
         Task <string> GetAccountNumberByUser(int userId);
-        Task <bool> WithdrawMoney(int userId, double amount);
-        Task <bool> DepositMoney(int merchantId, double amount);
+        Task <bool> WithdrawMoney(int userId, decimal amount);
+        Task <bool> DepositMoney(int merchantId, decimal amount);
     }
 }
